A lot job with design and logic of code
Have been done a lot job with design. Design logic is close to finish result.
Have been done a lot of work with program logic and structure. It became more versatile and easier to expand.
Added handler to 'click on bar' event .
Added animation of filling gradient of bar, from 0 to the current percentage of bar.
Added calculation of the percentage of the week.
Code get refracted and provided with comments.
Favicon is added.